requirements.txt
apache-airflow==2.3.4
psycopg2-binary==2.9.3
mysql-connector-python==8.0.31
pandas==1.5.2
sqlalchemy==1.4.46
